+++ src/components/loading.vue
@@ -1,0 +1,13 @@
+<template>
+  <img class="loading" src="@/static/media/loading.152225138e7c914c7d3f.gif" />
+</template>
+
+<style scoped lang="scss">
+img {
+  width: 40px;
+  height: 40px;
+  position: absolute;
+  top: calc(50% - 20px);
+  left: calc(50% - 20px);
+}
+</style>
